Extract 'prefix bundle'

A 'prefix bundle' is a string that starts with a numeric digit and is followed by 1 or 2 alpha characters. The string is passed to internal read_words function to parse out 'bundle' alphas from 'non-bundle' alphas found in the term. Returns the "bundle prefix" with no underscore prefix.

Usage

gather_n_move_prefix_num_bundle(x, relo_2_end = T, sep = "_")

Arguments

x

a character vector where matches are sought, or an object which can be coerced by as.character to a character vector. Long vectors are supported.

relo_2_end

logical, if TRUE then the prefix bundle, if it exists will be scooted to the end of the string x, else it will receive the necessary "_" prefix

sep

string of only one character long, intended to separate words from one another, not a starting character if original term has a numeric prefix!

Details

Currently, this function accounts for the following types of prefix bundles, where 3 is used as an example starting digit:

x = "3a. hey" will return "3a"

x = "3_17a_hey" will return "3_17a"

x = "3_17_hey" will return "3_17"

x = "3_17hey" will return "3_17"

x = "3hey" will return "3"

Examples

gather_n_move_prefix_num_bundle(x = "iAteABunch_of_grapesUntil99%Full")
#> Error in gather_n_move_prefix_num_bundle(x = "iAteABunch_of_grapesUntil99%Full"): could not find function "gather_n_move_prefix_num_bundle"
gather_n_move_prefix_num_bundle(x = "1a. How Was Brunch?")
#> Error in gather_n_move_prefix_num_bundle(x = "1a. How Was Brunch?"): could not find function "gather_n_move_prefix_num_bundle"
gather_n_move_prefix_num_bundle(x = "2aa How Was Brunch2day?")
#> Error in gather_n_move_prefix_num_bundle(x = "2aa How Was Brunch2day?"): could not find function "gather_n_move_prefix_num_bundle"
gather_n_move_prefix_num_bundle(x = "30abHow Was Brunch2day?")
#> Error in gather_n_move_prefix_num_bundle(x = "30abHow Was Brunch2day?"): could not find function "gather_n_move_prefix_num_bundle"
gather_n_move_prefix_num_bundle(x = c("30ab.How Was Brunch2day?", "40AHow Was Brunch2day?"))
#> Error in gather_n_move_prefix_num_bundle(x = c("30ab.How Was Brunch2day?",     "40AHow Was Brunch2day?")): could not find function "gather_n_move_prefix_num_bundle"
gather_n_move_prefix_num_bundle(x = c("iAteABunch_of_grapesUntil99%Full","30ab.How Was Brunch2day?", "40AHow Was Brunch2day?"))
#> Error in gather_n_move_prefix_num_bundle(x = c("iAteABunch_of_grapesUntil99%Full",     "30ab.How Was Brunch2day?", "40AHow Was Brunch2day?")): could not find function "gather_n_move_prefix_num_bundle"