Suggest a new renamed / abbreviated term

Take multiple character vectors as inputs and return the final suggestion or the transfermation method used.

Usage

least_pushy_rename_method(
  char_len,
  relo_2_end = TRUE,
  sep = "_",
  original_varname,
  dict_varname,
  use_bundle,
  adj_orig,
  stem,
  abbr_stem,
  abbr_parsed
)

Arguments

char_len

the maximum char length the final suggestion can be

original_varname

a character vector

dict_varname

a character vector to use first if a dictionary term exists

adj_orig

the original name, adjusted for any percent symbols (%) or grapes

stem

the stem or root word(s), char

abbr_stem

the character abbreviation of the stem or root

abbr_parsed

the character abbreviation of the adj_orig

Examples