Rename terms for submission compliance from vector

The function takes a single character vector as input and converts each string to be submission compliant. This means the variable must be a maximum of 8 characters, ASCII only, and should contain only uppercase letters, numbers and must start with a letter. No other symbols or special characters should be included in these names. However, legacy studies started on or before December 17, 2016 may use the underscore character "_". This function is slightly more flexible than the submission criteria would allow, so use the arguments wisely. tidy_varnames performs the same logic, but directly renames the columns of a data.frame plus enforces more strict adherence to the regulatory guidelines mentioned above.

Usage

tidy_rename(
  original_varname,
  char_len = 8,
  relo_2_end = TRUE,
  letter_for_num_prefix = "x",
  sep = "",
  replace_vec = c(`'` = "", `"` = "", `%` = "_pct_", `#` = "_nmbr_"),
  dict_dat = data.frame(original_varname = character(), dict_varname = character()),
  letter_case = "upper",
  case = "parsed",
  return_df = FALSE,
  verbose = getOption("xportr.type_verbose", "none")
)

Arguments

original_varname

a character vector needing renaming

char_len

integer, the max number of characters allowed in the renamed strings

relo_2_end

logical, if TRUE: numerical prefix bundles will be relocated to the end of the string. A prefix bundle is determined as a number or "grouping of numbers separated by special characters and other punctuation". See details section for more info on prefix bundles.

letter_for_num_prefix

character. Will be ignored if relo_2_end = TRUE. Per CDISC & regulatory body requirements, variable names cannot start with a number, so if you want to leave the number of the left-hand side, use this argument to insert a starting letter before the numeric prefix.

sep

string of only one character long, intended to separate words from one another. In general, only an empty string ("") and underscore ("") are possible, but the user can insert a letter if desired. Note that "" is discouraged because it is only permissible for legacy studies started on or before Dec 17, 2016.

replace_vec

A named character vector where the name is replaced by the value. This is helpful for finding and replacing smaller sub-strings embedded in variable names and automatically converting them to a preferred phrase. For example: '%' -> "pct".

dict_dat

a data frame containing two variables: the original_varname and the dict_varname used to find and replace entire variable names with new ones. For example, "Subject ID" -> "SUBJID".

letter_case

character, with choices c("upper", "lower", "asis") allowing user to make the final renamed terms uppercase (the default), lowercase, or leave them as-is, respectively. Note, lowercase is discourage as it is not submission compliant.

case

character, see to_any_case for more info on alternate cases but some popular choices are "snake", "lower_camel", or "parsed" (the default). From the documentation, the "parsed" case parses out substrings and surrounds them with an underscore. Underscores at the start and end are trimmed. No lower or upper case pattern from the input string are changed. If sep = "", then underscores will be trimmed later.

return_df

logical, defaults to TRUE where entire dataset is returned from suggestion process, else just the suggestion column itself

verbose

logical

Details

Abbreviating variable names in the xportr pkg uses a step-by-step process detailed below. Each original variable name will be renamed based on the method performing the least amount of work to achieve the min character length requirement (8 chars), thus maintaining maximum originality. Therefore, the function only moves on to the next method on an as-needed basis; namely, when the term is still greater than 8 characters, even after lite modification.

Blanks: Any columns that are missing a variable name (i.e., the header was blank in the source file) will be renamed to 'V' plus the column position. So if the 2nd column is blank, it will receive a 'V2' rename.

Use dictionary of controlled terminology: For example: 'Subject ID' may better be suited as 'SUBJID' within your organization. Note, that dictionary terms are expected to be submission compliant and will not be further abbreviated. They will, however, undergo a check for non-compliance.

Do nothing! Or at the very least, mimic what SAS does automatically when cleaning up variable names during a PROC IMPORT. Namely, replace any special characters & capitalize everything, etc. If the 'SASified' name is <= 8 chars, then the function will use that rename. However, if its still too long, the function will try removing any extra special characters or spaces to help reduce to 8 chars.

Find the STEM or ROOT word of each original variable name. For example, if the original contains the word 'resting', the 'ing' will be dropped and only the root word 'rest' will be considered. If less than 8 chars, the algorithm suggests that result. If its still too long, the function will, again, remove any special characters or spaces from the stemmed word(s).

Apply an abbreviation algorithm who's primary goal is readability, such that the results remain unique. The methods described below are a bit more 'involved', but the results are very robust. First, you should know that characters are always stripped from the end of the strings first (i.e. from right to left). If an element of the variable name contains more than one word (words are separated by spaces) then at least one letter from each word will be retained.

Method: First spaces at the ends of the string are stripped. Then (if necessary) any other spaces are stripped. Next, lower case vowels are removed followed by lower case consonants. Finally if the abbreviation is still longer than 8 chars, upper case letters and symbols are stripped.

When identifying 'words', the app performs some pre-processing steps in the following order:

Certain symbols are replaced. Like the '%' symbol is replaced with 'PCT'. Ex: 'I_Ate_%' becomes 'I_Ate_PCT'

Replace any symbols with a blank space. Ex: 'I_Ate' becomes 'I Ate'

Find when there are two capital letters next to each other, followed by a lower case letter then adds a space between the two capital letters to separate the assumed 'words'. Ex: 'IAte' becomes 'I Ate'

Insert a space between a number followed by a character. Ex: iAte1meal' becomes 'iAte1 meal'

Insert a space between a character followed by a number. Ex: 'iAte1meal' becomes 'iAte 1meal'

What do we abbreviate when? If a stemmed word exists, the app will apply the abbreviation algorithm described above on the stemmed version of the variable name, else the original variable name.

Since submission guidelines indicate variables may not start with a number, when found, the algorithm will either add and maintain a '_' "prefix bundle" through any transformations detailed above, or it will relocate the "prefix bundle" to the end of the term (the default behavior). What if a term starts with non-standard numerical prefix? Currently, the function accounts for the following types of prefix bundles, where 3 is used as an example starting digit:

x = "3a. hey" will return "3a"

x = "3_17a_hey" will return "3_17a"

x = "3_17_hey" will return "3_17"

x = "3_17hey" will return "3_17"

x = "3hey" will return "3"

See also

Examples

vars <- c("", "studyid", "STUDYID", "subject id", "1c. ENT", "1b. Eyes",
         "1d. Lungs", "1e. Heart", "year number", "1a. Skin_Desc")

# Default behavior
tidy_rename(vars)
#> 
#> The following variable name validation checks failed:
#> * Must be 8 characters or less: Variables `subject id`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> * Must start with a letter: Variables `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, and `1a. Skin_Desc`.
#> * Cannot contain any non-ASCII, symbol or underscore characters: Variables `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> * Cannot contain any lowercase characters Variables ``, `studyid`, `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> 
#> 
#> ── 10 of 10 (100%) variables were renamed ──
#> 
#> Var 1:  '' was renamed to 'V1'
#> Var 2:  'studyid' was renamed to 'STUDYID'
#> Var 3:  'STUDYID' was renamed to 'STUDYID2'
#> Var 4:  'subject id' was renamed to 'SUBJECTD'
#> Var 5:  '1c. ENT' was renamed to 'ENT1C'
#> Var 6:  '1b. Eyes' was renamed to 'EYES1B'
#> Var 7:  '1d. Lungs' was renamed to 'LUNGS1D'
#> Var 8:  '1e. Heart' was renamed to 'HEART1E'
#> Var 9:  'year number' was renamed to 'YEARNUMB'
#> Var 10:  '1a. Skin_Desc' was renamed to 'SKNDSC1A'
#> 
#> All renamed variables passed validation.
#> 
#>  [1] "V1"       "STUDYID"  "STUDYID2" "SUBJECTD" "ENT1C"    "EYES1B"  
#>  [7] "LUNGS1D"  "HEART1E"  "YEARNUMB" "SKNDSC1A"
tidy_rename(vars, letter_case = "asis")
#> 
#> The following variable name validation checks failed:
#> * Must be 8 characters or less: Variables `subject id`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> * Must start with a letter: Variables `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, and `1a. Skin_Desc`.
#> * Cannot contain any non-ASCII, symbol or underscore characters: Variables `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> * Cannot contain any lowercase characters Variables ``, `studyid`, `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> 
#> ── 8 of 10 (80%) variables were renamed ──
#> 
#> Var 1:  '' was renamed to 'V1'
#> Var 4:  'subject id' was renamed to 'subjectd'
#> Var 5:  '1c. ENT' was renamed to 'ENT1c'
#> Var 6:  '1b. Eyes' was renamed to 'Eyes1b'
#> Var 7:  '1d. Lungs' was renamed to 'Lungs1d'
#> Var 8:  '1e. Heart' was renamed to 'Heart1e'
#> Var 9:  'year number' was renamed to 'yearnumb'
#> Var 10:  '1a. Skin_Desc' was renamed to 'SknDsc1a'
#> 
#> The following variable name validation checks still failed:
#> Cannot contain any lowercase characters Variables `studyid`, `subjectd`, `ENT1c`, `Eyes1b`, `Lungs1d`, `Heart1e`, `yearnumb`, and `SknDsc1a`.
#>  [1] "V1"       "studyid"  "STUDYID"  "subjectd" "ENT1c"    "Eyes1b"  
#>  [7] "Lungs1d"  "Heart1e"  "yearnumb" "SknDsc1a"

# Leave numerical prefix on left-hand side, but add a starting letter
tidy_rename(vars, relo_2_end = FALSE, letter_for_num_prefix = "x")
#> 
#> The following variable name validation checks failed:
#> * Must be 8 characters or less: Variables `subject id`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> * Must start with a letter: Variables `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, and `1a. Skin_Desc`.
#> * Cannot contain any non-ASCII, symbol or underscore characters: Variables `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> * Cannot contain any lowercase characters Variables ``, `studyid`, `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> 
#> ── 10 of 10 (100%) variables were renamed ──
#> 
#> Var 1:  '' was renamed to 'V1'
#> Var 2:  'studyid' was renamed to 'STUDYID'
#> Var 3:  'STUDYID' was renamed to 'STUDYID2'
#> Var 4:  'subject id' was renamed to 'SUBJECTD'
#> Var 5:  '1c. ENT' was renamed to 'X1CENT'
#> Var 6:  '1b. Eyes' was renamed to 'X1BEYES'
#> Var 7:  '1d. Lungs' was renamed to 'X1DLUNGS'
#> Var 8:  '1e. Heart' was renamed to 'X1EHEART'
#> Var 9:  'year number' was renamed to 'YEARNUMB'
#> Var 10:  '1a. Skin_Desc' was renamed to 'X1SKNDSC'
#> 
#> All renamed variables passed validation.
#> 
#>  [1] "V1"       "STUDYID"  "STUDYID2" "SUBJECTD" "X1CENT"   "X1BEYES" 
#>  [7] "X1DLUNGS" "X1EHEART" "YEARNUMB" "X1SKNDSC"

# Add a dictionary and remove underscores
tidy_rename(vars, sep = "",
 dict_dat = data.frame(original_varname = "subject id", dict_varname = "subjid"))
#> 
#> The following variable name validation checks failed:
#> * Must be 8 characters or less: Variables `subject id`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> * Must start with a letter: Variables `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, and `1a. Skin_Desc`.
#> * Cannot contain any non-ASCII, symbol or underscore characters: Variables `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> * Cannot contain any lowercase characters Variables ``, `studyid`, `subject id`, `1c. ENT`, `1b. Eyes`, `1d. Lungs`, `1e. Heart`, `year number`, and `1a. Skin_Desc`.
#> 
#> ── 10 of 10 (100%) variables were renamed ──
#> 
#> Var 1:  '' was renamed to 'V1'
#> Var 2:  'studyid' was renamed to 'STUDYID'
#> Var 3:  'STUDYID' was renamed to 'STUDYID2'
#> Var 4:  'subject id' was renamed to 'SUBJID'
#> Var 5:  '1c. ENT' was renamed to 'ENT1C'
#> Var 6:  '1b. Eyes' was renamed to 'EYES1B'
#> Var 7:  '1d. Lungs' was renamed to 'LUNGS1D'
#> Var 8:  '1e. Heart' was renamed to 'HEART1E'
#> Var 9:  'year number' was renamed to 'YEARNUMB'
#> Var 10:  '1a. Skin_Desc' was renamed to 'SKNDSC1A'
#> 
#> All renamed variables passed validation.
#> 
#>  [1] "V1"       "STUDYID"  "STUDYID2" "SUBJID"   "ENT1C"    "EYES1B"  
#>  [7] "LUNGS1D"  "HEART1E"  "YEARNUMB" "SKNDSC1A"