Suggest a new renamed / abbreviated term

Take multiple character vectors as inputs and return the final suggestion or the transfermation method used.

Usage

least_pushy_rename_method(
  char_len,
  relo_2_end = TRUE,
  sep = "_",
  original_varname,
  dict_varname,
  use_bundle,
  adj_orig,
  stem,
  abbr_stem,
  abbr_parsed
)

Arguments

char_len

the maximum char length the final suggestion can be

original_varname

a character vector

dict_varname

a character vector to use first if a dictionary term exists

adj_orig

the original name, adjusted for any percent symbols (%) or grapes

stem

the stem or root word(s), char

abbr_stem

the character abbreviation of the stem or root

abbr_parsed

the character abbreviation of the adj_orig

Examples

least_pushy_rename_method(
  char_len = 8,
  original_varname = c("", "subject id", "1c. ENT", "1b. Eyes", "1d. Lungs",
     "1e. Heart", "year number", "1a. Skin_Desc"),
  dict_varname = c(NA, "SUBJID", NA, NA, NA, NA, NA, NA),
  use_bundle = c("","","","","","","",""),
  adj_orig = c("", "subject id", "1c. ENT", "1b. Eyes", "1d. Lungs",
     "1e. Heart", "year number", "1a. Skin_Desc"),
  stem = c("", "subject id", "c ent", "b eye", "d lung", "e heart",
     "year number", "a skin desc"),
  abbrev = c("", "subjctid", "_1c. ENT", "_1b.Eyes", "_1d lung", "_1eheart",
     "yearnmbr", "_1asknds"),
  abbr_transf = c("","","","","","","")
  )
#> Error in least_pushy_rename_method(char_len = 8, original_varname = c("",     "subject id", "1c. ENT", "1b. Eyes", "1d. Lungs", "1e. Heart",     "year number", "1a. Skin_Desc"), dict_varname = c(NA, "SUBJID",     NA, NA, NA, NA, NA, NA), use_bundle = c("", "", "", "", "",     "", "", ""), adj_orig = c("", "subject id", "1c. ENT", "1b. Eyes",     "1d. Lungs", "1e. Heart", "year number", "1a. Skin_Desc"),     stem = c("", "subject id", "c ent", "b eye", "d lung", "e heart",         "year number", "a skin desc"), abbrev = c("", "subjctid",         "_1c. ENT", "_1b.Eyes", "_1d lung", "_1eheart", "yearnmbr",         "_1asknds"), abbr_transf = c("", "", "", "", "", "",         "")): could not find function "least_pushy_rename_method"